cmon

+++ list_size.c
@@ -1,8 +1,8 @@
 #include "push_swap.h"
 
-size_t	list_size(t_n *s)
+int	list_size(t_n *s)
 {
-	size_t	i;
+	int	i;
 
 	i = 0;
 	while (s)
